refactor(rating-point): widen value type from byte to int

+++ backend/services/rating-service/rating/src/main/java/vn/uit/edu/msshop/rating/application/dto/query/RatingView.java
@@ -12,5 +12,5 @@ public record RatingView(
         UUID userId,
         String userName,
         String userAvatar,
-        byte ratingPoint) {
+        int ratingPoint) {
 }
